community-server/requirements.txt
Flask
flask-cors
google-generativeai
requests
beautifulsoup4
edge-tts
pyautogui
